make informer-gen respect the resourceName tag
